controller/requirements.txt
Django>=1.11
mysqlclient
psycopg2
pandas
requests


